Added source files
The source files/modules have been under development for past few days.
There are no tests for the files and cannot guarantee that they are
working.

Main aim of the project is to create bruteforce library that is fast
enough to break into system. It should be able to work with html
forms, files with password or anything involving bruteforce attack.

To improve performance, threads and asyncio were used to help increase
the number of attempts made in second. Asyncio part is running much
faster than threads part even asyncio is running on single thread.

source files description
------------------------

attempt.py - Responsible for communicating with target or system.
check.py - Analyses responce or output of attemp to check if access was granted
attack.py - Combination of both attack.py and check.
          - Communicates with target/system and also analyse responce
          - It only single attempt just as attempt.py
bruteforce.py - (to be renamed) Performs multiple attack with multiple attempts.
              - Uses threads and asyncio to improve performance
              - Realise that its classes may be renamed in future

handle_data.py - Responsible for preparing and hadling data for attack(e.g usernames, passwords)
               - Converts input data e.g 'array of passwords' into way that can be worked on easily.
               - This may include finding combination(e.g combination of usernames and passwords)
               - Theres a lot of work needed in this module

credentials.py - Responsible representing data to be sent to target/system.
	       - 'credential' correspond to data to be sent to target e.g username and password
	       - 'credential' may also be 'login_data'

column.py - (currently empty) Represents a column of data
	  - Let me say it represents things like 'passwords' or 'usernames'
	  - I cant come up with a good name, maybe saying 'column' is too bad

util.py - Contains commom methods to be used by other modules

Project idea
------------

The project was started as result of need to perform bruteforce attack on html form.
Tried searching the web for any Python library that does do it but couldnt find library
but examples on how to do it.

The project scope is not to be limited to html form or attack on website login pages
but to work on any situation. Its the consumers of the library that will define the
interface to that particular system/target.

For performing attack on webpage form login, the library user will have to use whatever Python
library for performing internet requests. By extending Attack or AttackAsync classes they may
be able to communicate with the system using whatever request library they used.

Luckily, the library comes with WebAttack and WebAttackAsync which will use request and
aiohttp to perform request in the target url and send the provided data(credentials). The only
thing to do would be implementing .success() method which should return True if system was unlocked.

Prgress
-------

The project seems likely to be successful based on the existing source files. Most low-level
activities have been implemented. Whats left is high level api that would simplify interraction
with the low-level functions.

project name
------------

Project name is 'Perock' which is likely to be name of the library.