Removed dependency on handle_data.py
Closes #1

Its not worth it to import 'handle_data' module in forcetable.py.
handle_data.py does not seem to be having future in this project
so its not worth for this project to depend on it.

credentials.py was the one depending on it but now credentials.py
will no longer used and will be removed as soon as issue #2 is
resolved.

handle_data.py may be assigned a new purpose such as reading data
from files. It wont be removed but the project wont be using the
module for now. Reading data from file may be out of scope of the
project but it may be worth it.