Import modules relatively(Closes #6)
Importing modules relatevely is important if perock is to be become
a package. Perock is expected to imported as package or maybe be dstributed
to pypi which would require the package in source/ to be importable. Without
relative import it becomes impossible to import the package without import
error being raised.

I have experienced this kind of issues but I have worked up on them by using
relative imports. Tried everything I could but it seems like relative import
is mandory when module is part of package.