Implemented BForce.consumer() use of semaphore
Previous BForce.consumer() implementation had to wait for items from produser
before starting tasks. It had to also wait for those tasks to complete before
starting new tasks.

Use of semaphore eliminated the need to have to wait which allows task to be exectuted
as soon as theres space for it. This is expected to improve performance of consumer
and also prevent it from being locked for long time when waiting for all tasks to
complete.

The need to wait for all was eliminated but it wasnt for BForceAsync.consumer().
No matter how much I tried to implement it, I could come up with a way it could work.

See issue #8 for more