Removed requests and aiohttp as dependencies(Closes #7)
The project shouldnt be depending on requests and aiohttp as they are just
used by few classes which arent even used in the project other than when
testing.

The two modules will be imported dynamically when then they are to be used
by the objects of their classes. Attack class imports requests and AttackAsync
imports aiohttp.

See issue #7 for more info.