WebAttack and WebAttack async moved to attacks.py
attacks.py file/modukle was created and the classes were added to it.
Existing import statement that were depending on the classes were also
modified to refer to attacks module instead of attack module.

Classes in attacks.py are meant to be used by users/consumers of perock
not the library itself. The classes may require certain libraries like
requests. It will be the consumers of the library responsibility to install
them.

This commit is related to closed issue #7.