Added TestAttack and TestAttackAsync classes
Both the classes are to be used for testing purposes. They
may be valuable when testing BForce and BForceAsync classes.
They implement waiting using .sleep() method of either time
or asyncio modules.

Now its not neccessay to always use WebAttack which requires internet
connection to work. The two test classes are reliable for testing purposes
as their completion times does not vary that much with retries.