Added primary column support and few methods
This commit is an efford to resolve issue #10. Having primary columns support
is whats also required to resolve the issue. The other added methods are for
accessing this primary columns.

Currently based on issue #10, only single primary column will be used but
FTable also support multiple primary columns. The last added primary column
is the one that becomes primary column for the Ftable object.