Added success behaviour on produser(Closes #10)
This commit is much closely related to the last 2 commits and are
all related to issue #10. Main issue was that produser and consumer
for BForce and BForceAsync were running until the end on all rows
even if corresponding rows were successful/hacked.

Produser and Consumer should be able to decide if certain rows are worth
it to try. Not all items needs to tried, but certain items that are already
tried should be able to influence if one similar row should be tried.

One implementation is using primary column. Rows with same primary column value
or item will only tried once. That means only one of the rows will be tried
with the other rows skipped.

Example, only a single instance of a username will be tried with the rest of username
and passwords combinations skipped. That would improve performance especially if it
takes long to communicate with target.

Advantages
----------
- Improved performance
- Reduses number of requests to target

Disadvantages
-------------
- Does not stop/remove existing items
- Consumer does not react immediately
- Continually loops every combination
- Its fast, but not that fast due to looping.

To remove bottlenecks, produser should create rows manually for each primary colunm items
and then should stop doing so as soon as theres success. That would preventhaving to create
unneccessay combinations and looping over them.

The commit was about success behaviou, but also fail behavious is neccessary.
What should happen when it fails or there are errors when communicating with target?