Fixed 'producer' spell errors(Closes #12)
'produser' was replaced by 'producer', this was a fix for spell error
for producer. All occurrances were replaced including functions, attributes
and on comments.

The spell errors wasnt not only on the project source code but also on issues
and commit messages. Commit messages and iessues comments wont be fixed as its not that
possible as it would result in rewritting history and its also not worth it and impractical.