Fixed issue #13
Closes #13

BForce produser was looping and checking through all FRows in FTable even if it
was obvious that the FRows wont be used to create tasks. A way to avoid having to check
and loop unneccessary FRows was neeeded. This commit attempts to fix the problem of
having to perform a task that could have been cut too early.

This commit introdused new method to help in solving the issue. 'producer(self)' method was
modified to call other methods instead of handling producer functionalities itself.
Two methods called by 'producer()' method were added which is 'producer_loop_all()' and
'producer_loop_some()'.

'producer()' calls one of methods depending on the state of the object. The other methods
introdused in this commit help in getting producer method to be called by 'producer()' method.
'producer_loop_all()' is the same as the original producer() but 'producer_loop_some()' is the
one that was just implemented which is the main focus of this commit.

'producer_loop_some()' only loops and check only neccessay FRows and it will stop looping
certain rows when corresponding FRow get successful or cracks the system. This is based
on primary column and its items. FRows with same primary column would only until one of
the FRows creacks the system.

If one FRow cracks the system, then other Frows with same primary column item will never be
tried. This is common for username-password attack, if username is cracked then theres no
need to check if other passwords combinations with the username cracks the system. That is
because most of the time theres single username in system.

Thats why 'producer_loop_some()' requires primary column in FTable object, if not it raises
exception.

Advantages
----------
- Its faster than previous implementation.
- Allows to choose producer method.

Disadvantages
-------------
- Requires primary column.
- No benefit on small FRows if not slow down.
- No benefits if primary column has duplicates.
- May take long if primary column is too large.
- Existing tasks arent stopped when switching to another primary item.

The last disadvanges has to do with not checking primary column item before trying to create rows.
Its also because producer does not stop stop existing running tasks which causes delay as it has
to wait for the tasks to complete before producer can pick the new new tasks from the new FRows.

The disadvantages would be smaller when FTable is larger or primary items dont contain duplicates.
The producer takes as if primary column items are unique so if it has duplicates there they will be
tried just as as other items in primary column. It wont run for long if its Frows should be used
to create tasks but would have taken some looping time.

Its better to ensure that FTable columns or FColumn dont contain duplicates. Theres no way to do it
as FColumn views its items as coming from iterator. Theres no way one can remove duplicates from iterator
without reading contents of iterator into memory or writting them somewhere. set() wont help as that would
write iterator into memory which is not worth as FColumn can contain millions of items.