Fixes and improvements to bruteforce.py
Changes
-------
- 'self.ftable_queue' was renamed to 'self.rows_queue'
- Added methods for cancelling producer and consumer.
- Added method for checking if consumer or producer shoud continue running.
- Removed timeout for getting frows in queue.
- Consumer for BForceAsync updated(no longer calls `ftable_queue_elements(self)`.

The main objection of this commit was to remove use of timeout when getting items from frows queue.
Its not known when item will be available and if frow is not available at moment that does not mean
it wont be available. Before this commit, consumer would stop starting new tasks when frow cannot
be found in frows queue.

It may be possible that producer may be slow to put items and consumer should just terminate. COnsumer
should wait for producer to first terminate before it terminates. It should check both if frows queue is
empty and if producer completed. If producer completed and there are no frows left, then its right time
to exit if not wait for producer to put the frows.

That is great as producer may load the data from source in which the time the data can arrive is not
known. There is no way consumer can know when producer will have data available, but consumer can wait
for producer to provide data.

Thats why producer-consumer or master-slave idea was created in the first place.