Fixed tests setUp()
The SetUP() method was made to be reusable by splitting it into small
methods. This allows subclass to overide those methods which make SetUp()
ethod more reusable. It also avaoid having to create a lot of objects from
super class which are not even used by sub classes.

Example, TestAttackCommon is subclass of TestAttemptCommon and TestCheckCommon.
SetUp of the two base classes creates objects used by them like Attempt and Check
objects. Some of the objects are not used by TestAttackCommon as using them would mean
its not testing the objects it was supposed to test.

TestAttackCommon should test Attack class not Attempt or Check but needs to reuse the methods
of TestAttemptCommon and TestCheckCommon. Thats why there a need for TestAttackCommon to be able
to have control on how TestAttemptCommon and TestCheckCommon setUp() creates some of objects.

Manually creating them would waste time and make code more complex or introduce bugs. Calling
certain methods withing setUp() of TestAttemptCommon and TestCheckCommon is the best option just
as its done with regular __init__() method.