Fixes and updates from tests
Changes and modifications in  this commit are results of writting tests
for 'target' and 'brutforce' modules.

Target class of 'target' module was updated to return  Responce object instead of string. '
test_attempt.py' was updated to refect changes to match changes to the new changes to
Target class to fix failing tests. The modules receuved minor changes which may be bug
fixes detected from tests, changes to make code more reusable or addition of new methods.

Tests for 'target' and 'brutforce' modules will be committed in the next commit.

Related to issue #4