Add: product.py
'product' module contains Product class and product() function for
calculating Cartesian product of iterables similar to itertools.product().
Problem is that itertools.product() raises MemoryError if one of iterables
is too large.

itertools.product() casts some of iterables into tuple which end up writting
the whole iterator into memory. Thats because python iterators are not reusable
after being iteratoed completely(exausted). itertools.product() fights that by
converting iterators to tuple which can be reused even after being iterated completely.

Generators also cannot be iteratoted again after being completely iterated. The module
fights that by accepting callables that will be called when the iterator is needed. The
callables will be the ones creating iterators when needed so the problem of iterators
being exausted is eliminated.

Creating callables for iterators seems to be adding complexity for code thats why
product() function of the module does call itertools.product() if iterables does
not contain iterators.

Advantages
----------
- No MemoryError
- Iterators are not written to mememory.
- Works for both iterator and callable.

Disadvantages
-------------
- Slower than itertools.product()
- RecursionError maybe raised if there so many callables(cuurently 12 or more)
- Larget repeat may raise RecursionError
- Requires callables instead of iterators which can cause confution.

Issues
------
- The module has complicated code which can make it harder to fix errors or maintain.
- Increasing number of collables or repeat can cause recursion error.
- Error will be raised if iterables contain both iterators and callables.
- Its harder to extend number of iterables that can supported without recursion error.
- There are lot code that should have been reused but was copy and pasted.

Currently repeat should be less than 12 and number of callables be less than 12 for it to
work without recursion error. It was once limited to 3 but after some tricks it increased to
5 and then by magic to 11.

It seems possible to increase it to infinite.
But why would someone want product of 12 or more iterators?