Bug Fix: Fixed unraised exception on producer()
producer() on BForceAsync was not raising errors and consumer
not stopping even if producer experienced exception. Problem was with
that consumer is waiting for producer to funish but producer was never
able to do so.

That was solved by hadling producer exceptions and also calling 'self.producer_done_callback()'
on finally block to alert that producer finished so that consumer should stop waiting
for producer.