Add: session.py
'session' contain classes that implement sessiion methods of BForce classes.
The module is not be used by the project but it may be worth it for reducing
complexity of 'bforce' module.

Currently, BForce classes contain session methods themselfs. 'session' module
is an ettempt to remove session methods from BForce classes.