Update: attacks.py, attempt.py, bruteforce.py
attempt classes were updated to be to set exceptions not to be raised
on .request() method. Previously, all exceptions were not raised but would
be set as responce.

Its only the implementation of Attack class that knows which exceptions not
to be raised. Catching all exceptions is a bad choice as it would make finding
bugs with .request(self) method harder.