Add: responce.py
'responce' module is meant for assisting with handling of responce to determine
if responce means success/cracked or attempt failed. Its only the users
who knows the outcomes of attempt so there is no way this module
can help know if attempt was success or failure.

One thing that this module can do is to simplify the proccess based on certain
output of responce. Most targets or systems returns would return text to client
like webpages, mobile apps, deskop apps and its through this text in which
users receive messages.

By analying the text from responce, one may be able to learn something from the
text. Text proccessing can be slow but checking if certain words exits in
text can fast. Thats what this module attempts to archieve which is to
help in deciding letting users focus on other features that matters most.

Target may not only communicate through text but other means may be used but
most of the time it will be text especially if target is not located locally.
Exceptions can also be responces like exception from failure to open file
with given password.

The module contains ResponceBytes class for representing responce text. It
makes it easy to check if subtring exists than ordinary str or bytes  objects.
The classs instances internally keeps bytes object but the classs itself can work
with strings as it internally convert strings to bytes.

Its better to use bytes as responce most of the time will be in bytes like that
one from responce from server. But client code will be using strings most of
the time. Thats why ResponceBytes class is compatible with both strings and
bytes.

The class did not implement __eq__() as I was afraid it would cause problems with
hashing(dict, sets). But the class instances are expected to be the same if they
have same text like b'text' and 'text'. No matter if its string or bytes they are
all considered the same.