Update: Added AttackText and AttackTextAsync classes
AttackText and AttackTextAsync classes are built on top of BytesCompare
class of 'responce' module. They may help in making it easy to analyse responce
in which responce could contain text or bytes.

The classes can be used for situation like web bruteforce attack in which responce
from server will be text like html. test_attack.py file was updated to include tests
for the two classes.