Add: consumer.py, producer.py Closes #15
Producer methods of 'bforce' module are now split into their own
module. This wasnt just moving the methods but whole new changes
with new classes with some new features.

Some of the changes include parallel primary items records being produced
as of issue #17. The commit will also make it possible to resolve isse #16
and allow producer to not run on its own worker(e.g thread).

This commit is also enough to close issue #17 but that issue will be resoved
when 'producer' module is integrated to 'bforce' module.