Implements parallel primary tasks Closes #17
Its now should be possible execute primary tasks as in issue #17.
This should allow multiple usernames to be tried on username-password attack.
It could be what someone expects in some situation or maybe improve performance.

Note that this only applied to loop some producer and not implemented for loop all producer.
Source code for how loop some implements that can be seen within 'producer' module or in previous commit.