Updates: attempt, check and attack classes
This commit was an efford to fix attack classes to make them easy to use.
To make that possible, other classes had to be modified as attempt and check
classes.

There were no major changes to high level methods other renaming of .start_request()
to .start(). This commit is expected to make it easier to use attack classes and make
bug fixes more easier as it now raises exceptions if outcomes are not in expected way.

Changes
-------
- request_error() and target_reached() are now part of Check classes.
- target_reached() is now coroutine(awaitable) on AttackAsync.
- request_started() was added to Check class.
- request_errors() was renamed to request_failed().

- Attack classes methods are now validated with _validate_attack().
- Attack class responce massages are now updated by _update_responce_message().

- Attack classes now support retry/running until target is reached.
- .start_request() was renamed to .start().
- .after_start_request() and .before_start_request() were renamed to .after_request() and .before_request().