Update: 'forcetable' module
'forcetable' module was updated to make it easy to create table.
The changes in this commit would make it easier to add support for
json or csv as table.

The main function is .records_to_table() which takes iterator of Records
objects and creates Table objects. Its so powerful in that it can create
original fields if primary_field_name is passed.

Changes
-------
- Added records_to_item_names_map(), records_to_fields() and records_to_table() functions.
- Added new class RecordsTable.
- Renamed MapTable to PrimaryTable.