Modify: Added Single Leading Underscore
Internal instance atrributes were had underscore added to them.
Its just for showing off that the attributes are meant to be used
within the object. Its just better for them to have a leading underscore
other than not having one.

It would be better if they are accessed by methods other being accessed directly
as it seems if the underscore is not added.