Refactor: bforce.py & runner.py closes #9, #16
closes #9, closes #16

bforce.py/'bforce' module was unneccessay complex and not reusable.
It caused lot of issues including slowness, dependency on threading objects.
Its classes were also harder to reuse and some classes inherited methods they never used.
It was great when it was small with only two classes but now it grew very large

Changes in this commit arent just for 'bforce' module, but 'runner' module and tests.
'runner' module was refactored as its based on 'bforce' module.

Changes
-------
bforce.py - Refactored with new classes added based on existing classes.
          - BForceBase class was added which acts as base class for all other bforce classes.
          - BForceParallel was added which is base class for concurrent/parallel bforce class e.g BForceAsync
	  - BForceExecutor class was added for bforce classes using executors(e.g ones using threadpoolexecutor)
          - BForceThread class was added for executing bruteforce with threads(implements BForceExecutor).

          - BForce class now inherit/implements BForceThread.
          - BForceBase class introdused updates to session methods, producer, consumer and other methods.
          - consumer() methods now uses handle_attack_recursive() as inspired by BForceAsync.

	  - producer is now used as iterator other than execute in its own worker(based on issue #16).
          - clear_queue(), handle_attacks() methods were removed.
          - to_future(), handle_attack_recursive_future() and wait_tasks() were added to BForceParallel.

          - BForceAsync now inherit BForceParallel which does not us ethreading objects(based on issue #9).
          - BForceBlock now inherit BForceBase and some of its methods removed as they implemented by base class.

          - There are lot other changes not mensioned.

runner.py - Refactored similar to bforce.py.
          - Original Runner classs was split to multiple including RunnerBase, RunnerParallel, etc.
          - _after_create_bforce_object() method added to RunnerBase.
          - Some portions of _create_bforce_object() have been split to _after_create_bforce_object().
          - _bforce attribute was added to RunnerBase and used by _after_create_bforce_object().
          - _after_request() was renamed to _after_attack().
          - _setup_before_object() was added to RunnerBase and .start() updated to use it.

Remaining files are test files for runner.py and bforce.py. They were updated to reflect recent changes
to their corresponding modules. The test file swere also refactored similar to their source files such
as adding new classes.

test_bforce.py within integrated tests wasnt changed as its tests methods werent failing or raising
exceptions. It would be worth it to update it, but it since it does not cause any problems its better
to leave and focus on other issues.

Based on itegrated tests, this commit makes things much faster as it took less time to execute the tests.
It can be though as 2X improve in speed which can be attributed to producer not using threading queue and
not running in its own thread(#16).

This commit also makes classes it modified more reusable including the test classes. That could make it easier
to support running bruteforce attack on multiple processes as it does not work based on issue #26. It also
opens opportunity to resolve issue #25 without having to depend on 'contextvars' module.

Final words: This commit took a lot of effords complete.
           : It was thought while in process of distributing the project and releasing the repository to public.
           : Licence file was added, but had to stop everything for case of making the project work with py36.