Removed use of f-string
f-string is not supported by python 3.5 which would result in syntax error
on versions of python less than 3.6. By removing any use of f-string, perock
will be able to work with python versions less than 3.6.

Removing f-string does not guarantee that perock will work with python 2.5 but
just increases changes for it to be. There may be other features perock is using
that were just added in python 3.6 especially those of asynio.

Note: attack.py is missing as its local history on my side was lost.
    : it was overided by vscode and now it looks as if its a new untracked file without refs.
    : committing it may overide its history in remote repository.