Update: README.md & setup.cfg
Perock does not seem to support python 3.4 due to use of type hints.
setup.cfg was updated to require Python 3.5 and later while README.md
was updated to show which versions it was tested including python 3.4 and less
not supported.

Changes
-------
setup.cfg - 'python_requires' was updated to python 3.5 and later.
README.md - Environment heading was added with python versions perock was tested.
          - 'Instagram Bruter' was added as similar project.
          - Perock 'cons' were updated.