Updated some files
This commit was intensionally for solving issue #26 which was failed
to be solved. Adding support for multiprocessing does not seem easy based
on the implementations of BForceExecutor.

All of the changes were just commented within bforce.py. Another thing is realisation
of that perock does not support python 3.5 which was in previous commits stated that
perock supports python 3.5 and later.

It was due to tox running system wide python when tesing which made me conclude that perock
worked well on the versions of python being tested. Another thing is that there are errors
when running tests with python 3.10 which are caused by aiounittest.

README.md and setup.cfg were updated to show that perock requires python version 3.6 and higher.
Even if there are errors within tests on python 3.10, it did perform well when manually tests.
Manual tests arent as reliable as automated tests but that proves that perock realy works on python 3.10.

Its been too long developing perock, its time to somehow release it.