Closes #28
There was a bug that caused some records not be tried even if optimise was not
enabled on runner if table contains single field. Ths commit fixes the bug to allow
all records to be tried when optimise is disabled.

There are also some minor changes for readability which are not related to issue #28.