Closes #31, 'prodius' and 'fortable' support.
Support for 'prodius' and 'fortable' libraries was added to perock.
The two libraries were developed based on this librarry(perock).
They are more matured and perform better than corresponding modules in perock.

Changes made were minimal as possible to avoid having to change many files.
The libraries are not compatiple with perock since the have lowercase classes.
Aliases were used to make the classes be imported as lowercase.

There are other things that made libraries incmpatible such as renamed functions.
Some portions had to be changed to match how that portion is defined in libraries.

Tests were also failing most of the time mostly integrated tests.
'forcetable' library had to be updated with new releases to fix certain bugs.
Currently all tests are doing well but there may be other compatibility issues.