Future API and Implementation(#36, #44)
Modules to be used for refactoring and changing perock API has been added.
They are similar to existing code but differ in that methods are static or class methods.
Methods are no longer tied to object such as 'success()'.

Modules added in this commit will change the way perock works.
Attack, bforce and runner classes will be removed after they are completed(decreased source code).
Existing modules are left to not break 'broote' library and fix bugs exposed to 'broote'.